initial version of dependent service plugins.
